no crash on crossings on final tiles
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@1287 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/components/gui_image.h
@@ -19,10 +19,10 @@ class gui_image_t : public gui_komponente_t
 {
 private:
 	image_id id;
-	uint8 player_nr;
+	uint16 player_nr;
 
 public:
-    gui_image_t(image_id i=IMG_LEER, uint16 p=0 ) { id = i; player_nr = p; }
+    gui_image_t(image_id i=IMG_LEER, uint8 p=0 ) { id = i; player_nr = p; }
 
     void set_image(image_id i) { id = i; }
